updated attributes and methods
I added methods and attributes to the Light Novel and Anime class. The instance information is displayed in a dictionary instead of a formatted string. The added attributes provides more information about the object of a particular class. The Manga class inherits from the light novel class. The Anime class is completely separate since the attributes and methods are not related to the Light Novel or Manga class.